+++ djangae/settings_base.py
@@ -56,4 +56,4 @@ DJANGAE_RUNSERVER_IGNORED_FILES_REGEXES = ['^.+$(?<!\.py)(?<!\.yaml)(?<!\.html)'
 # Note that these should match a directory name, not directory path:
 DJANGAE_RUNSERVER_IGNORED_DIR_REGEXES = [r"^google_appengine$"]
 
-TEST_RUNNER = 'djangae.test_runner.DjangaeDiscoverRunner'
+TEST_RUNNER = 'djangae.test.DjangaeDiscoverRunner'
