+++ djangae/test.py
@@ -4,6 +4,8 @@ import os
 
 from django import test
 from django.test import Client
+from django.test.runner import DiscoverRunner
+from djangae.test_runner import bed_wrap
 
 from djangae.environment import get_application_root
 
@@ -171,3 +173,10 @@ class TestCase(HandlerAssertionsMixin, TestCaseMixin, test.TestCase):
 
 class TransactionTestCase(HandlerAssertionsMixin, TestCaseMixin, test.TransactionTestCase):
     pass
+
+
+class DjangaeDiscoverRunner(DiscoverRunner):
+    def build_suite(self, *args, **kwargs):
+        suite = super(DjangaeDiscoverRunner, self).build_suite(*args, **kwargs)
+        suite._tests[:] = [bed_wrap(test) for test in suite._tests]
+        return suite
