+++ docs/installation.md
@@ -37,6 +37,7 @@ if __name__ == "__main__":
     from djangae.core.management import execute_from_command_line, test_execute_from_command_line
 
     if "test" in sys.argv:
+        # This prevents the local sandbox initializing when running tests
         test_execute_from_command_line(sys.argv)
     else:
         execute_from_command_line(sys.argv)
